refactor: rename libscarlet to scarlet_std and implement a bump allocator re-export: core, alloc crate as scarlet_std
