+++ user/lib/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "libscarlet"
+name = "scarlet_std"
 version = "0.8.1"
 edition = "2024"
 
 [lib]
-name = "scarlet"
+name = "scarlet_std"
 path = "src/lib.rs"
 
 [build]
