+++ user/lib/src/lib.rs
@@ -1,18 +1,91 @@
+//! # Scarlet Standard Library
+//!
+//! This no_std library provides the core functionality for user-space programs 
+//! running on the Scarlet..
+//!
+//! ## Features
+//!
+//! - Re-exports of core and alloc library components for use in a no_std environment
+//! - System call interface for interacting with the Scarlet kernel
+//! - Architecture-specific functionality
+//! - Custom memory allocator implementation
+//!
+//! ## Usage
+//!
+//! This library is automatically linked to user-space programs in the Scarlet.
+//! It provides the essential functionality needed to interact with the kernel
+//! and manage resources in a user-space environment.
+//!
+//! The library is deliberately minimal to ensure small binary sizes and fast
+//! compile times while still providing the necessary functionality for
+//! application development.
 #![no_std]
 #![no_main]
 #![feature(naked_functions)]
 #![feature(alloc_error_handler)]
+#![feature(async_iterator)]
+#![feature(new_range_api)]
 
-extern crate alloc;
+mod core_exports {
+    extern crate core;
+
+    pub use core::any;
+    pub use core::array;
+    pub use core::async_iter;
+    pub use core::cell;
+    pub use core::char;
+    pub use core::clone;
+    pub use core::cmp;
+    pub use core::convert;
+    pub use core::default;
+    pub use core::future;
+    pub use core::hint;
+    pub use core::i8;
+    pub use core::i16;
+    pub use core::i32;
+    pub use core::i64;
+    pub use core::i128;
+    pub use core::isize;
+    pub use core::iter;
+    pub use core::marker;
+    pub use core::mem;
+    pub use core::ops;
+    pub use core::option;
+    pub use core::pin;
+    pub use core::ptr;
+    pub use core::range;
+    pub use core::result;
+    pub use core::u8;
+    pub use core::u16;
+    pub use core::u32;
+    pub use core::u64;
+    pub use core::u128;
+    pub use core::usize;
+}
+
+mod alloc_exports {
+    extern crate alloc;
+
+    pub use alloc::borrow;
+    pub use alloc::boxed;
+    pub use alloc::collections;
+    pub use alloc::fmt;
+    pub use alloc::format;
+    pub use alloc::rc;
+    pub use alloc::slice;
+    pub use alloc::str;
+    pub use alloc::string;
+    pub use alloc::vec;
+}
 
 mod arch;
-pub mod mem;
+mod allocator;
 pub mod syscall;
+pub use core_exports::*;
+pub use alloc_exports::*;
 
 #[panic_handler]
 pub fn panic(_info: &core::panic::PanicInfo) -> ! {
-    // This is the panic handler.
-    // You can put your code here.
     loop {}
 }
